Refactor translation function
So I rewrite the translation js. I was able to convert your IFs code to a neat json object thanks to sublime. The old version takes a lot of memory and its super slow because of the redundant IF`s. Well you can't notice if you have an high end pc lol. But here is a better approach, les work. All translations are inside a JSON object. You can easily add new translation with this, more convenient.